backend/deepfake_detector/requirements.txt
Django==5.1.7
djangorestframework==3.16.0
numpy==2.2.4
Pillow==11.2.0
tensorflow==2.19.0
tensorflow_hub==0.16.1
